lambda/requirements.txt
google-cloud-compute==1.22.0
google-auth==2.22.0
boto3>=1.26.0
requests>=2.31.0
